Compute neighborhood statistics from raster
This is the latest python notebook for computing neighborhood statistics from the DOGAMI rasters.  It still obtains the geometries (shapes) from the postgis database, but I will try to get a version that can also take shape files directly.  It now has the capability of classifying the geometry statistics by binning the raster statistics using pandas cut function.  Also, ALL parameters are now specified in a single data dictionary at the top for easy conversion to a json config file.