Create disaster_neighborhoods.sql
SQL to create the disaster_neighborhoods table. This is populated with data extracted from the DOGAMI rasters by Fred Pearce.  The source file is a CSV file called DisasterNeighborhoodsFixed_DogamiRasters_stats.
I used the pgAdmin import function to load the data into the production database, so there's no associated SQL for that.